requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
langchain-community
psycopg2-binary
